Do not redirect to the wizard if network activate

+++ init.php
@@ -185,10 +185,12 @@ register_activation_hook( __FILE__, 'it_exchange_activation_hook' );
  * @return void
 */
 function it_exchange_register_activation_hook() {
-	if ( false !== get_option( '_it-exchange-register-activation-hook', false ) ) {
-        delete_option('_it-exchange-register-activation-hook');
-		wp_safe_redirect('admin.php?page=it-exchange-setup' );
-    }
+	if ( ! is_network_admin() ) {
+		if ( false !== get_option( '_it-exchange-register-activation-hook', false ) ) {
+		    delete_option('_it-exchange-register-activation-hook');
+			wp_safe_redirect('admin.php?page=it-exchange-setup' );
+		}
+	}
 }
 add_action( 'admin_init', 'it_exchange_register_activation_hook' );
